removed mapster
