+++ sample/Sample.Restaurant.Server/Infrastructure/MappingConfiguration.cs
@@ -1,0 +1,17 @@
+using Mapster;
+using MapsterMapper;
+
+namespace Sample.Restaurant.Server.Configuration;
+
+internal static class MappingConfiguration
+{
+    public static TypeAdapterConfig AddMapper(this IServiceCollection services)
+    {
+        services.AddTransient<IMapper, ServiceMapper>();
+
+        var config = new TypeAdapterConfig();
+        services.AddSingleton(config);
+
+        return config;
+    }
+}
