+++ src/Lewee.Application/ApplicationConfiguration.cs
@@ -1,8 +1,6 @@
 ﻿using System.Reflection;
 using FluentValidation;
 using Lewee.Application.Mediation.Behaviors;
-using Mapster;
-using MapsterMapper;
 using MediatR;
 using MediatR.Pipeline;
 using Microsoft.Extensions.DependencyInjection;
@@ -62,20 +60,5 @@ public static class ApplicationConfiguration
         services.AddTransient(typeof(IPipelineBehavior<,>), typeof(RequestPostProcessorBehavior<,>));
 
         return services;
-    }
-
-    /// <summary>
-    /// Adds type mapper
-    /// </summary>
-    /// <param name="services">Service collection</param>
-    /// <returns>Mapster type addapter config</returns>
-    public static TypeAdapterConfig AddMapper(this IServiceCollection services)
-    {
-        services.AddTransient<IMapper, ServiceMapper>();
-
-        var config = new TypeAdapterConfig();
-        services.AddSingleton(config);
-
-        return config;
     }
 }
