+++ src/Lewee.Application/Mediation/Requests/IQuery.cs
@@ -8,7 +8,7 @@ namespace Lewee.Application.Mediation.Requests;
 /// <typeparam name="T">
 /// Query response type
 /// </typeparam>
-public interface IQuery<T> : IApplicationRequest, IRequest<T>
+public interface IQuery<T> : IApplicationRequest, IRequest<QueryResult<T>>
     where T : class
 {
 }
